更名，从ProductTableCell更改为ShopsTableCell

+++ BinFenV10/AboutMeViewController.m
@@ -116,7 +116,7 @@ typedef enum
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    int sectionIndex = indexPath.section;
+    NSInteger sectionIndex = indexPath.section;
     switch (sectionIndex)
     {
         case SectionIndexUser:
